попытки в модульность как у фольксвагена
