+++ js/main.js
@@ -1,98 +1,4 @@
-//Функция возвращающая случайное целое число число из заданного диапозона(включительно)
-
-const getRandomNumber = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  const errorMessage = ('Wrong value, try again');
-  if (max <= min) {
-    return errorMessage;
-  }
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-
-//Функция возвращающая случайное дробное число из заданных параметров включительно. Количество знаков после запятой = 5
-
-const getRandomFloat = (min, max) => {
-  const errorMessage = ('Wrong value, try again');
-  if (max <= min) {
-    return errorMessage;
-  }
-  const getRandomFloat = Math.random() * (max - min + 1) + min;
-  return +getRandomFloat.toFixed(5);
-}
-
-//Функция возвращающая случайный элемент массива
-
-const getRandomElement = (array) => {
-  return array[getRandomNumber(0, array.length - 1)];
-};
-
-//Функция возвращающая массив из рандомных элементов заданного масива (Рандомайзер)
-
-const getRandomArray = (array) => {
-  const numberArray = getRandomNumber(0, array.length - 1);
-  const randomArray = [];
-  for (let i = 0, l = numberArray; i < l; i++) {
-    const arrayElement = getRandomNumber(0, array.length - 1);
-    if (!randomArray.includes(array[arrayElement])) {
-      randomArray.push(array[arrayElement]);
-    }
-  }
-  return randomArray;
-};
-
-//Набор данных и массивов для получения "объявлений о сдаче"
-
-const AvatarValue = { MIN: 1, MAX: 8 };
-const TITLE_DATA = 'Информация о помещении.';
-const GeoDataValueX = { MIN: 35.65000, MAX: 35.70000 };
-const GeoDataValueY = { MIN: 139.70000, MAX: 139.80000 };
-const MAX_PRICE_VALUE = 20000;
-const placeType = ['palace', 'flat', 'house', 'bungalow'];
-const MAX_ROOMS_VALUE = 4;
-const MAX_GUESTS_VALUE = 10;
-const checkinTime = ['12:00', '13:00', '14:00'];
-const checkoutTime = ['2:00', '13:00', '14:00'];
-const featuresData = ['elevator', 'dishwasher', 'parking', 'wifi', 'washer', 'conditioner'];
-const DESCRIPTION_DATA = 'Неплохое место что бы отдохнуть Вам и Вашим близким';
-const photosData = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
-const MAX_ARRAY_VALUE = 10;
-
-
-//Функция возвращающая массив из рандомных данных
-
-const getArrayRandomData = () => {
-  const getRandomAvatarNumb = getRandomNumber(AvatarValue.MIN, AvatarValue.MAX);
-  const getGeoDataX = getRandomFloat(GeoDataValueX.MIN, GeoDataValueX.MAX);
-  const getGeoDataY = getRandomFloat(GeoDataValueY.MIN, GeoDataValueY.MAX);
-
-  return {
-    author: {
-      avatar: 'img/avatars/user0' + getRandomAvatarNumb + '.png',
-    },
-    offer: {
-      title: TITLE_DATA,
-      address: {
-        x: getGeoDataX,
-        y: getGeoDataY
-      },
-      price: getRandomNumber(1, MAX_PRICE_VALUE),
-      type: getRandomElement(placeType),
-      rooms: getRandomNumber(1, MAX_ROOMS_VALUE),
-      guests: getRandomNumber(1, MAX_GUESTS_VALUE),
-      checkin: getRandomElement(checkinTime),
-      checkout: getRandomElement(checkoutTime),
-      features: getRandomArray(featuresData),
-      description: DESCRIPTION_DATA,
-      photos: getRandomArray(photosData),
-    },
-    location: {
-      x: getGeoDataX,
-      y: getGeoDataY
-    },
-  };
-};
+import {getArrayRandomData, MAX_ARRAY_VALUE} from './data.js';
 
 //Функция возвращающая массив из полученых рандомных данных
 
@@ -102,3 +8,4 @@ const createArrayDeclarations = (arrayCount) => {
 
 const declarations = createArrayDeclarations(MAX_ARRAY_VALUE);
 
+console.log(declarations);
